app/requirements.txt
alembic==1.6.2
astroid==2.9.2
attrs==21.4.0
black==21.12b0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
grpcio==1.37.1
grpcio-tools==1.37.1
idna==2.10
isort==5.10.1
itsdangerous==2.0.0
Jinja2==3.0.3
jsonschema==4.3.3
lazy-object-proxy==1.7.1
ldap3==2.9
Mako==1.1.4
MarkupSafe==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
Paste==3.5.0
pathspec==0.9.0
platformdirs==2.4.1
protobuf==3.17.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
pyOpenSSL==20.0.1
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
sqlalchemy-json==0.4.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
waitress==2.0.0
Werkzeug==1.0.1
wrapt==1.13.3
